Version 0.1
This is the first stable verison of the webstie to be tested in
TU-Berlin subnetwork. This version is the basic layout and functionality
for the website to be created. This version does not have any control to
take measurements on the test bench. It is just setup to test the
website running on a server in the subnetwork of MDT (Chair of
Electronic Measurement and Diagnostic Technology) without failures.

Furthermore this version is missing the following pages in the navbar:

- Nutzung

- Webcam