Fix css and js files to work with current html files
To get the site working properly the .js scripts are now received
throuhg internet. The css file is saved locally. However, it is intended
to get the website working with only local css and js files to be able
to test the website while beeing offline. Furthermore the website could
be used even if there is no external internet available. Maybe this will
lead to performance improvements as well.